+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Williams](https://togaware.com/Graham.Williams.html).
 
 ## 1.0 Functional App
 
++ Drag stations into your own order on the Stations screen [1.0.4 20260922 gjw]
 + Add support for macOS installer builds [1.0.3 20260922 gjw]
 + Generated all icons [1.0.2 20260921 gjw]
 + Add snapcraft config to build a snap [1.0.1 20260921 gjw]